# goforms/contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GoFormX API",
    "version": "1.2.0",
    "license": {
      "name": "GNU Affero General Public License v3.0 or later",
      "identifier": "AGPL-3.0-or-later"
    },
    "description": "The control plane manages forms and immutable schema versions. The public\ndata plane exposes published schemas and accepts validated submissions.\nSchema and submission JSON numbers preserve exact numeric values. Clients\nmust not round numbers through binary floating point before sending them or\nafter reading them. Numeric tokens are limited to 4096 bytes, explicit\nexponents to -1024 through 1024, and expanded values to 1024 integer and\n1024 fractional decimal places. Excessive numeric representations return\n400 rather than being rounded. JSONB may normalize numeric notation;\nnumerically equivalent retries remain idempotent.\nEvery requestBody operation requires exactly its documented media type.\nContent-Type may omit parameters or include only `charset=utf-8`\n(case-insensitive); missing headers, other media types, charsets, or\nparameters return 415 `unsupported_media_type`. Request documents must be\none JSON object. Duplicate object member names are rejected before decoding\nat every nesting depth, including escape-equivalent names such as `a` and\n`\\u0061`; malformed or trailing JSON returns 400 `invalid_request`.\n"
  },
...
  "servers": [
    {
      "url": "https://api.goformx.com",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/v1/forms/{formId}/submissions/export": {
      "parameters": [
        {
          "$ref": "#/components/parameters/FormId"
        }
      ],
      "post": {
        "summary": "Prepare a bounded redacted submission download and durable audit record",
        "operationId": "exportSubmissions",
        "x-goformx-export-limits": {
          "rows": 1000,
          "sourceBytes": 8388608,
          "outputBytes": 8388608,
          "processingSeconds": 10,
          "csvColumns": 256,
          "csvDepth": 32,
          "requestBytes": 4096,
          "concurrentExports": 1
        },
        "tags": [
          "Control plane"
        ],
        "security": [
          {
            "serviceToken": []
          },
          {
            "firstPartyAssertion": []
          }
        ],
        "x-goformx-required-scopes": [
          "submissions:read"
        ],
        "description": "Exports one owned, non-deleted form from a single database statement\nsnapshot, newest received time and ID first. Filters belong in the JSON\nbody, never a query string. Unknown, repeated, null, and empty fields\nare rejected. Body limit is 4096 bytes; there is no cursor or limit override.\nHard bounds are 1000 rows, 8 MiB of source payload/policy/metadata, 8 MiB\nof encoded output, and a 10-second application processing deadline.\nExceeding a bound releases no partial download. The accepted version's\nprivacy policy applies identically to normal reads. No reveal mode exists.\nBefore returning any download bytes, the service persists an append-only\nexport.prepared audit containing actor, organization, form, credential\nclass/non-secret ID, request ID, format, row/byte counts and time, but no\npayloads or filters. Prepared does not prove client receipt.\nJSON preserves exact numeric values. CSV flattens nested objects to\nJSON-Pointer columns; arrays/empty objects remain JSON cell text. CSV is\nlimited to 256 total columns and 32 nested object levels. Every cell,\nincluding headers and numbers, is double-quoted, quote-escaped and\napostrophe-prefixed as spreadsheet text. This is an initial-download\ndefense, not a guarantee after arbitrary spreadsheet edits/re-saves.\nOne export at a time is admitted per API instance, without a wait queue;\nconcurrent authorized requests receive 429 with Retry-After.\nUse JSON for machine round trips, distinguishing missing/empty cells,\nand exact numeric processing. Responses are attachments with no-store.\nContent-Length declares the complete artifact size; download proxies must\nverify it against received bytes before offering the file to a browser.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "format"
                ],
                "properties": {
                  "format": {
                    "type": "string",
                    "enum": [
                      "json",
                      "csv"
                    ]
                  },
                  "receivedFrom": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Inclusive lower bound; explicit offset and at most microsecond precision."
                  },
                  "receivedBefore": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Exclusive upper bound; must be later than receivedFrom."
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "accepted"
                    ]
                  },
                  "schemaVersion": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 2147483647
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Complete bounded export prepared and durably audited",
            "headers": {
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 8388608
                }
              },
              "X-GoFormX-Export-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "maxItems": 1000,
                      "items": {
                        "$ref": "#/components/schemas/Submission"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "exportId",
                        "preparedAt",
                        "rowCount"
                      ],
                      "properties": {
                        "exportId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "preparedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "rowCount": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 1000
                        }
                      }
                    }
                  }
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "413": {
            "description": "Row, source-byte, output-byte, CSV depth, or column budget exceeded; narrow filters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "429": {
            "description": "Another export is active on this API instance; retry later",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Durable audit failed; no download was released",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Export processing deadline exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "serviceToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "gfst_ opaque token",
        "description": "External integration credential stored server-side, returned once, and verified by hash, organization owner, lifecycle, and required operation scope. Browsers must not receive it.",
        "x-goformx-credential-class": "external-service-token"
      },
      "firstPartyAssertion": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT (EdDSA)",
        "description": "Single-use, at-most-60-second assertion minted by the Waaseyaa server after session authentication and organization membership resolution. Never delivered to a browser and never accepted as an external service token.",
        "x-goformx-credential-class": "first-party-assertion",
        "x-goformx-jwt-profile": "gofx-fpa-v1",
        "x-goformx-jwks-uri": "https://goformx.com/.well-known/goformx-control-plane-jwks.json"
      }
    },
    "parameters": {
...
      "FormId": {
        "name": "formId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "schemas": {
...
      "Submission": {
        "type": "object",
        "required": [
          "id",
          "formId",
          "schemaVersion",
          "requestId",
          "status",
          "data",
          "submittedAt",
          "redactedPaths"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "formId": {
            "type": "string",
            "format": "uuid"
          },
          "schemaVersion": {
            "type": "integer",
            "minimum": 1
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "description": "Correlation identifier for the accepted public request; it contains no credential or payload data."
          },
          "status": {
            "type": "string",
            "enum": [
              "accepted"
            ],
            "description": "Immutable acceptance state. Delivery attempts have their own lifecycle."
          },
          "data": {
            "type": "object",
            "additionalProperties": true,
            "description": "Redacted projection of the immutable payload. Marked object members are omitted, marked array elements are null, and a root pointer yields an empty object. Never use this projection for validation or resubmission."
          },
          "redactedPaths": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Sorted JSON Pointers declared sensitive by the accepted schema, including absent optional targets. Empty when no fields are marked. Values are never returned via a reveal flag."
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "FieldError": {
        "type": "object",
        "required": [
          "pointer",
          "code",
          "message"
        ],
        "properties": {
          "pointer": {
            "type": "string",
            "description": "JSON Pointer into the request document"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "fields": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                }
              }
            }
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Stable error envelope; status is determined by the HTTP response code",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "UnsupportedMediaType": {
        "description": "Content-Type is missing or does not match the operation's documented JSON media type and optional UTF-8 charset policy",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "unsupported_media_type",
                "message": "Content-Type must match the operation's documented JSON media type; only an optional UTF-8 charset parameter is accepted.",
                "requestId": "req_01K4Q9X2JXQG67A4WZ8X9CEQ4G"
              }
            }
          }
        }
      },
...
    }
  },
...
}